# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/link/": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "link"
                ],
                "summary": "Creates a short link",
                "parameters": [
                    {
                        "type": "string",
                        "format": "url",
                        "description": "url to shorten",
                        "name": "url",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handler.createLinkResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.JSONError"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/handler.JSONError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.JSONError"
                        }
                    }
                }
            }
        },
        "/api/link/{link_id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "link"
                ],
                "summary": "Get information about shortened link",
                "parameters": [
                    {
                        "type": "string",
                        "format": "url",
                        "description": "Link ID",
                        "name": "link_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.getLinkInfoResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.JSONError"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/handler.JSONError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.JSONError"
                        }
                    }
                }
            }
        },
        "/l/{link_id}": {
            "get": {
                "tags": [
                    "link"
                ],
                "summary": "Redirect to source url",
                "parameters": [
                    {
                        "type": "string",
                        "description": "link ID",
                        "name": "link_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "308": {
                        "description": "Permanent Redirect"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handler.JSONError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "handler.createLinkResponse": {
            "type": "object",
            "properties": {
                "link_id": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "handler.getLinkInfoResponse": {
            "type": "object",
            "properties": {
                "url": {
                    "type": "string"
                }
            }
        }
    }
}